+++ tests/testthat/test-satres.R
@@ -3,7 +3,7 @@ test_that("satres()", {
     esa <- system.file("extdata", "esa/f", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = TRUE)
+                 only_spectral_bands = TRUE)
     c(
       names(sr$bands),
       names(sr$bands[["r1000m"]]),
@@ -75,7 +75,7 @@ test_that("satres()", {
     esa <- system.file("extdata", "esa/f", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = FALSE)
+                 only_spectral_bands = FALSE)
     c(
       names(sr$bands),
       names(sr$bands[["r1000m"]]),
@@ -214,7 +214,7 @@ test_that("satres()", {
     esa <- system.file("extdata", "esa", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = TRUE)
+                 only_spectral_bands = TRUE)
     c(
       names(sr$bands),
       names(sr$bands[["r1000m"]]),
@@ -285,7 +285,7 @@ test_that("satres()", {
     esa <- system.file("extdata", "esa", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = FALSE)
+                 only_spectral_bands = FALSE)
     c(
       names(sr$bands),
       names(sr$bands[["r1000m"]]),
@@ -425,7 +425,7 @@ test_that("get_spatial_resolution()", {
     esa <- system.file("extdata", "esa", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = TRUE)
+                 only_spectral_bands = TRUE)
     sr |>
       get_spatial_resolution()
   },
@@ -433,29 +433,137 @@ test_that("get_spatial_resolution()", {
 })
 
 
-test_that("as_SpatRaster()", {
+test_that("get_band_names()", {
   expect_equal({
     esa <- system.file("extdata", "esa", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = TRUE)
-    r <- sr |>
-      as_SpatRaster("r1000m")
-    names(r)
+                 only_spectral_bands = FALSE)
+    sr |>
+      get_band_names()
   },
-  c("B02", "B03", "B04", "B08"))
+  c(
+    "AOT",
+    "B01",
+    "B02",
+    "B03",
+    "B04",
+    "B05",
+    "B06",
+    "B07",
+    "B08",
+    "B09",
+    "B11",
+    "B12",
+    "B8A",
+    "MSK_CLDPRB",
+    "MSK_DETFOO_B01",
+    "MSK_DETFOO_B02",
+    "MSK_DETFOO_B03",
+    "MSK_DETFOO_B04",
+    "MSK_DETFOO_B05",
+    "MSK_DETFOO_B06",
+    "MSK_DETFOO_B07",
+    "MSK_DETFOO_B08",
+    "MSK_DETFOO_B09",
+    "MSK_DETFOO_B10",
+    "MSK_DETFOO_B11",
+    "MSK_DETFOO_B12",
+    "MSK_DETFOO_B8A",
+    "MSK_SNWPRB",
+    "SCL",
+    "WVP"
+  ))
+})
+
+test_that("get_band_names()", {
+  expect_equal({
+    esa <- system.file("extdata", "esa", package = "satres")
+    sr <- satres(dir = esa,
+                 out_dir = tempdir(),
+                 only_spectral_bands = FALSE)
+    sr |>
+      get_band_names("r1000m")
+  },
+  c(
+    "AOT",
+    "B02",
+    "B03",
+    "B04",
+    "B08",
+    "MSK_DETFOO_B02",
+    "MSK_DETFOO_B03",
+    "MSK_DETFOO_B04",
+    "MSK_DETFOO_B08",
+    "WVP"
+  ))
 })
 
 
-test_that("save_by_resolution()", {
+test_that("get_band_names()", {
+  expect_equal({
+    esa <- system.file("extdata", "esa", package = "satres")
+    sr <- satres(
+      dir = esa,
+      out_dir = tempdir(),
+      only_spectral_bands = TRUE
+    )
+    sr |>
+      get_band_names(c("r1000m", "r2000m"))
+  },
+  c(
+    "B01",
+    "B02",
+    "B03",
+    "B04",
+    "B05",
+    "B06",
+    "B07",
+    "B08",
+    "B11",
+    "B12",
+    "B8A"
+  ))
+})
+
+test_that("get_spectral_band_names()", {
   expect_equal({
     esa <- system.file("extdata", "esa", package = "satres")
     sr <- satres(dir = esa,
                  out_dir = tempdir(),
-                 only_bands = TRUE)
-    r <- sr |>
-      save_by_resolution(only_show_files = TRUE)
-    basename(r)
+                 only_spectral_bands = FALSE)
+    sr |>
+      get_spectral_band_names()
   },
-  c("r1000m.tif", "r2000m.tif", "r6000m.tif"))
+  c(
+    "B01",
+    "B02",
+    "B03",
+    "B04",
+    "B05",
+    "B06",
+    "B07",
+    "B08",
+    "B09",
+    "B11",
+    "B12",
+    "B8A"
+  ))
+})
+
+test_that("get_spectral_band_names()", {
+  expect_equal({
+    esa <- system.file("extdata", "esa", package = "satres")
+    sr <- satres(dir = esa,
+                 out_dir = tempdir(),
+                 only_spectral_bands = FALSE)
+    sr |>
+      get_spectral_band_names("r1000m")
+  },
+  c(
+    "B02",
+    "B03",
+    "B04",
+    "B08"
+  ))
 })